refactor(auth): split AuthService and AuthLocalService
Extract AuthLocalService into its own module and keep AuthService as the base contract. This prepares the auth layer for additional provider implementations (GitHub, GitLab, Bitbucket) while preserving current local auth behavior and tests.
